refactor: Result 패키지 이동
